+++ src/internals/neighborlist.jl
@@ -40,13 +40,13 @@ function reduce_output!(output::NeighborList, output_threaded::Vector{<:Neighbor
 end
 
 # Estimate the expected number of neighbor pairs assuming uniform density.
-function _estimated_n_pairs(box::Box{UnitCellType, N}, cl::CellList) where {UnitCellType, N}
+function _estimated_n_pairs(box::Box{<:UnitCellType, N}, cl::CellList) where {UnitCellType, N}
     n = cl.n_real_particles
     vol = abs(LinearAlgebra.det(box.input_unit_cell.matrix))
     sphere_vol = (N == 2 ? π : 4π / 3) * box.cutoff^N
     return max(0, round(Int, (n * (n - 1) / 2) * sphere_vol / vol))
 end
-function _estimated_n_pairs(box::Box{UnitCellType, N}, cl::CellListPair) where {UnitCellType, N}
+function _estimated_n_pairs(box::Box{<:UnitCellType, N}, cl::CellListPair) where {UnitCellType, N}
     n_x = cl.ref_list.n_real_particles
     n_y = cl.target_list.n_real_particles
     vol = abs(LinearAlgebra.det(box.input_unit_cell.matrix))