+++ CHANGELOG.md
@@ -34,6 +34,7 @@ Version 0.10.0-DEV
 - ![INFO][badge-info] Internal implementation of neighborlist moved to ParticleSystem interface.
 - ![INFO][badge-info] The internal representation of coordinantes is done with `ParticleSytemPositions` array type - which is not a subtype of abstract array. 
 - ![INFO][badge-info] Add performance test.
+- ![INFO][badge-info] Default name for output in ParticleSystem is `default_output_name`, but it can be accessed though `sys.output`, as before.
 - ![BUGFIX][badge-bugfix] Fix cross-computation (`pairwise!(f, x, sys)`) with `NonPeriodicCell` failing to find pairs when particle coordinates span negative values or large coordinate ranges / do not modify input coordinates in `NonPeriodicCell`. This bug was never released, it was created and fixed in the development version.
 
 Version 0.9.17